--- cpp/src/gandiva/gdv_arithmetic_function_stubs.cc
@@ -1,56 +1,0 @@
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-//#pragma once
-
-#include "gandiva/engine.h"
-#include "gandiva/exported_funcs.h"
-#include "gandiva/formatting_utils.h"
-#include "gandiva/gdv_function_stubs.h"
-#include "gandiva/random_generator_holder.h"
-
-extern "C" {
-
-double gdv_fn_random(int64_t ptr) {
-  gandiva::RandomGeneratorHolder* holder =
-      reinterpret_cast<gandiva::RandomGeneratorHolder*>(ptr);
-  return (*holder)();
-}
-
-double gdv_fn_random_with_seed(int64_t ptr, int32_t seed, bool seed_validity) {
-  gandiva::RandomGeneratorHolder* holder =
-      reinterpret_cast<gandiva::RandomGeneratorHolder*>(ptr);
-  return (*holder)();
-}
-}
-
-namespace gandiva {
-
-void ExportedArithmeticFunctions::AddMappings(Engine* engine) const {
-  std::vector<llvm::Type*> args;
-  auto types = engine->types();
-
-  // gdv_fn_random
-  args = {types->i64_type()};
-  engine->AddGlobalMappingForFunc("gdv_fn_random", types->double_type(), args,
-                                  reinterpret_cast<void*>(gdv_fn_random));
-
-  args = {types->i64_type(), types->i32_type(), types->i1_type()};
-  engine->AddGlobalMappingForFunc("gdv_fn_random_with_seed", types->double_type(), args,
-                                  reinterpret_cast<void*>(gdv_fn_random_with_seed));
-}
-}  // namespace gandiva
