+++ cpp/src/gandiva/precompiled/CMakeLists.txt
@@ -118,7 +118,6 @@ if(ARROW_BUILD_TESTS)
                  ../gdv_time_function_stubs.cc
                  ../gdv_string_function_stubs.cc
                  ../gdv_hash_function_stubs.cc
-                 ../gdv_arithmetic_function_stubs.cc
                  ../../arrow/vendored/datetime/tz.cpp
                  hash_test.cc
                  hash.cc
